Add timestamp to service node seeding
This changes SeedServiceNode to also include an addedTimestamp field so
that when seeding we can backdate a node's actual oxen registration date
into the contract.

Seeding this fixes a bug where the early contributor leave delay would
apply for the first 30 days of the contract, even for a node that has
been registered long beyond 30 days.

It also has a (smaller) cosmetic benefit of having the service node
longevity tracked in the contract.